# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Krutrim Agent Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/providers/{agent_key}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "List Settings",
        "operationId": "list_settings_api_providers__agent_key__get",
        "parameters": [
          {
            "name": "agent_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Settings Api Providers  Agent Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{agent_key}/{role}": {
      "get": {
        "tags": [
          "providers"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_providers__agent_key___role__get",
        "parameters": [
          {
            "name": "agent_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Key"
            }
          },
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Settings Api Providers  Agent Key   Role  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "providers"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_providers__agent_key___role__put",
        "parameters": [
          {
            "name": "agent_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Key"
            }
          },
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/providers/{agent_key}/{role}/reset": {
      "post": {
        "tags": [
          "providers"
        ],
        "summary": "Reset Settings",
        "operationId": "reset_settings_api_providers__agent_key___role__reset_post",
        "parameters": [
          {
            "name": "agent_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Key"
            }
          },
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateSettingsResponse": {
        "properties": {
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "settings",
          "note"
        ],
        "title": "UpdateSettingsResponse",
        "description": "`settings` stays a free-form dict, not a typed `ModelSettings`\nsubmodel \u2014 the actual object is always a provider-specific subclass\n(e.g. `OpenRouterModelSettings`), and typing this as the generic base\nwould make FastAPI silently strip subclass-only fields from the\nresponse, not just the docs."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}